requirements.txt
goslate==1.5.4
jsonlines==3.1.0
keras
Keras_Preprocessing==1.1.2
langcodes==3.3.0
matplotlib
nltk==3.7
numpy==1.23.4
pandas==1.5.1
PyYAML==6.0
scikit_learn==1.1.2
seaborn
seqeval==1.2.2
spacy==3.4.2
tensorflow==2.10.0
torch==1.12.1
tqdm==4.64.1
transformers==4.23.1
translators==5.4.3
